requirements.txt
streamlit
pandas
azure-storage-blob 
azure-identity
plotly
python-dotenv
kaleido
